+++ MCP_Server/__init__.py
@@ -3,4 +3,4 @@
 __version__ = "0.1.0"
 
 # Expose key classes and functions for easier imports
-from .server import AbletonConnection, get_ableton_connection
+from .core import AbletonConnection, get_ableton_connection
