+++ MCP_Server/tools/playback_tools.py
@@ -1,0 +1,30 @@
+"""Playback control tools for Ableton MCP."""
+
+import logging
+from mcp.server.fastmcp import Context
+
+from ..core import get_ableton_connection
+
+logger = logging.getLogger("AbletonMCPServer")
+
+
+def start_playback(ctx: Context) -> str:
+    """Start playing the Ableton session."""
+    try:
+        ableton = get_ableton_connection()
+        result = ableton.send_command("start_playback")
+        return "Started playback"
+    except Exception as e:
+        logger.error(f"Error starting playback: {str(e)}")
+        return f"Error starting playback: {str(e)}"
+
+
+def stop_playback(ctx: Context) -> str:
+    """Stop playing the Ableton session."""
+    try:
+        ableton = get_ableton_connection()
+        result = ableton.send_command("stop_playback")
+        return "Stopped playback"
+    except Exception as e:
+        logger.error(f"Error stopping playback: {str(e)}")
+        return f"Error stopping playback: {str(e)}"
